requirements.txt
streamlit
openai>=1.0.0
scikit-learn
matplotlib
numpy
pandas
tiktoken
beautifulsoup4
requests
python-dotenv\n# Models configurable with VC_HUNTER_CHAT_MODEL and VC_HUNTER_EMBED_MODEL